--- frontend/web/components/organisation-settings/usage/UsageDashboard/index.ts
@@ -1,3 +1,0 @@
-export { default } from './UsageDashboard'
-export { default as UsageDashboardView } from './UsageDashboardView'
-export type { UsageDashboardViewProps } from './UsageDashboardView'
