+++ frontend/web/components/pages/usage/index.ts
@@ -1,0 +1,3 @@
+export { default } from './UsageDashboardPage'
+export { default as UsageDashboard } from './UsageDashboard'
+export type { UsageDashboardProps } from './UsageDashboard'
